+++ server/sonar-server/src/main/java/org/sonar/server/computation/CeQueue.java
@@ -20,16 +20,7 @@
 package org.sonar.server.computation;
 
 import com.google.common.base.Optional;
-import org.sonar.api.server.ServerSide;
-import org.sonar.api.utils.System2;
-import org.sonar.core.util.UuidFactory;
-import org.sonar.db.DbClient;
-import org.sonar.db.DbSession;
 import org.sonar.db.ce.CeActivityDto;
-import org.sonar.db.ce.CeQueueDto;
-import org.sonar.server.computation.monitoring.CEQueueStatus;
-
-import static java.lang.String.format;
 
 /**
  * Queue of pending Compute Engine tasks. Both producer and consumer actions
@@ -38,96 +29,40 @@ import static java.lang.String.format;
  *   This class is decoupled from the regular task type {@link org.sonar.db.ce.CeTaskTypes#REPORT}.
  * </p>
  */
-@ServerSide
-public class CeQueue {
+public interface CeQueue {
+  /**
+   * Build an instance of {@link CeTaskSubmit} required for {@link #submit(CeTaskSubmit)}. It allows
+   * to enforce that task ids are generated by the queue. It's used also for having access
+   * to the id before submitting the task to the queue.
+   */
+  CeTaskSubmit prepareSubmit();
 
-  private final System2 system2;
-  private final DbClient dbClient;
-  private final UuidFactory uuidFactory;
-  private final CEQueueStatus queueStatus;
-  private final CeQueueListener[] listeners;
+  /**
+   * Submits a task to the queue. The task is processed asynchronously.
+   * If submits are paused (see {@link #isSubmitPaused()}, then an
+   * unchecked exception is thrown.
+   */
+  CeTask submit(CeTaskSubmit submit);
 
-  // state
-  private boolean submitPaused = false;
-  private boolean peekPaused = false;
+  /**
+   * Peek the oldest task in status {@link org.sonar.db.ce.CeQueueDto.Status#PENDING}.
+   * The task status is changed to {@link org.sonar.db.ce.CeQueueDto.Status#IN_PROGRESS}.
+   * Does not return anything if the queue is paused (see {@link #isPeekPaused()}.
+   *
+   * <p>Only a single task can be peeked by project.</p>
+   *
+   * <p>An unchecked exception may be thrown on technical errors (db connection, ...).</p>
+   */
+  Optional<CeTask> peek();
 
-  public CeQueue(System2 system2, DbClient dbClient, UuidFactory uuidFactory,
-    CEQueueStatus queueStatus, CeQueueListener[] listeners) {
-    this.system2 = system2;
-    this.dbClient = dbClient;
-    this.uuidFactory = uuidFactory;
-    this.queueStatus = queueStatus;
-    this.listeners = listeners;
-  }
-
-  public CeTaskSubmit prepareSubmit() {
-    return new CeTaskSubmit(uuidFactory.create());
-  }
-
-  public CeTask submit(CeTaskSubmit submit) {
-    if (submitPaused) {
-      throw new IllegalStateException("Compute Engine does not currently accept new tasks");
-    }
-    CeTask task = new CeTask(submit);
-    DbSession dbSession = dbClient.openSession(false);
-    try {
-      CeQueueDto dto = new CeQueueDto();
-      dto.setUuid(task.getUuid());
-      dto.setTaskType(task.getType());
-      dto.setComponentUuid(task.getComponentUuid());
-      dto.setStatus(CeQueueDto.Status.PENDING);
-      dto.setSubmitterLogin(task.getSubmitterLogin());
-      dto.setStartedAt(null);
-      dbClient.ceQueueDao().insert(dbSession, dto);
-      dbSession.commit();
-      queueStatus.addReceived();
-      return task;
-    } finally {
-      dbClient.closeSession(dbSession);
-    }
-  }
-
-  public Optional<CeTask> peek() {
-    if (peekPaused) {
-      return Optional.absent();
-    }
-    DbSession dbSession = dbClient.openSession(false);
-    try {
-      Optional<CeQueueDto> dto = dbClient.ceQueueDao().peek(dbSession);
-      if (!dto.isPresent()) {
-        return Optional.absent();
-      }
-      queueStatus.addInProgress();
-      return Optional.of(new CeTask(dto.get()));
-
-    } finally {
-      dbClient.closeSession(dbSession);
-    }
-  }
-
-  public boolean cancel(String taskUuid) {
-    DbSession dbSession = dbClient.openSession(false);
-    try {
-      Optional<CeQueueDto> queueDto = dbClient.ceQueueDao().selectByUuid(dbSession, taskUuid);
-      if (queueDto.isPresent()) {
-        if (!queueDto.get().getStatus().equals(CeQueueDto.Status.PENDING)) {
-          throw new IllegalStateException(String.format("Task is in progress and can't be cancelled [uuid=%s]", taskUuid));
-        }
-        cancel(dbSession, queueDto.get());
-        return true;
-      }
-      return false;
-    } finally {
-      dbClient.closeSession(dbSession);
-    }
-  }
-
-  void cancel(DbSession dbSession, CeQueueDto q) {
-    CeActivityDto activityDto = new CeActivityDto(q);
-    activityDto.setStatus(CeActivityDto.Status.CANCELED);
-    remove(dbSession, new CeTask(q), q, activityDto);
-  }
-
+  /**
+   * Cancels a task in status {@link org.sonar.db.ce.CeQueueDto.Status#PENDING}. An unchecked
+   * exception is thrown if the status is not {@link org.sonar.db.ce.CeQueueDto.Status#PENDING}.
+   * The method does nothing and returns {@code false} if the task does not exist.
+   *
+   * @return true if the task exists and is successfully canceled.
+   */
+  boolean cancel(String taskUuid);
 
   /**
    * Removes all the tasks from the queue, whatever their status. They are marked
@@ -137,9 +72,7 @@ public class CeQueue {
    *
    * @return the number of canceled tasks
    */
-  public int clear() {
-    return cancelAll(true);
-  }
+  int clear();
 
   /**
    * Similar as {@link #clear()}, except that the tasks with status
@@ -148,83 +81,25 @@ public class CeQueue {
    *
    * @return the number of canceled tasks
    */
-  public int cancelAll() {
-    return cancelAll(false);
-  }
+  int cancelAll();
 
-  private int cancelAll(boolean includeInProgress) {
-    int count = 0;
-    DbSession dbSession = dbClient.openSession(false);
-    try {
-      for (CeQueueDto queueDto : dbClient.ceQueueDao().selectAllInAscOrder(dbSession)) {
-        if (includeInProgress || !queueDto.getStatus().equals(CeQueueDto.Status.IN_PROGRESS)) {
-          cancel(dbSession, queueDto);
-          count++;
-        }
-      }
-      return count;
-    } finally {
-      dbClient.closeSession(dbSession);
-    }
-  }
+  /**
+   * Removes a task from the queue and registers it to past activities. This method
+   * is called by Compute Engine workers when task is processed.
+   *
+   * <p>An unchecked exception is thrown if the task does not exist in the queue.</p>
+   */
+  void remove(CeTask task, CeActivityDto.Status status);
 
-  public void remove(CeTask task, CeActivityDto.Status status) {
-    DbSession dbSession = dbClient.openSession(false);
-    try {
-      Optional<CeQueueDto> queueDto = dbClient.ceQueueDao().selectByUuid(dbSession, task.getUuid());
-      if (!queueDto.isPresent()) {
-        throw new IllegalStateException(format("Task does not exist anymore: %s", task));
-      }
-      CeActivityDto activityDto = new CeActivityDto(queueDto.get());
-      activityDto.setStatus(status);
-      Long startedAt = activityDto.getStartedAt();
-      if (startedAt != null) {
-        activityDto.setFinishedAt(system2.now());
-        long executionTime = activityDto.getFinishedAt() - startedAt;
-        activityDto.setExecutionTimeMs(executionTime);
-        if (status == CeActivityDto.Status.SUCCESS) {
-          queueStatus.addSuccess(executionTime);
-        } else {
-          queueStatus.addError(executionTime);
-        }
-      }
-      remove(dbSession, task, queueDto.get(), activityDto);
+  void pauseSubmit();
 
-    } finally {
-      dbClient.closeSession(dbSession);
-    }
-  }
+  void resumeSubmit();
 
-  private void remove(DbSession dbSession, CeTask task, CeQueueDto queueDto, CeActivityDto activityDto) {
-    dbClient.ceActivityDao().insert(dbSession, activityDto);
-    dbClient.ceQueueDao().deleteByUuid(dbSession, queueDto.getUuid());
-    dbSession.commit();
-    for (CeQueueListener listener : listeners) {
-      listener.onRemoved(task, activityDto.getStatus());
-    }
-  }
+  boolean isSubmitPaused();
 
-  public void pauseSubmit() {
-    this.submitPaused = true;
-  }
+  void pausePeek();
 
-  public void resumeSubmit() {
-    this.submitPaused = false;
-  }
+  void resumePeek();
 
-  public boolean isSubmitPaused() {
-    return submitPaused;
-  }
-
-  public void pausePeek() {
-    this.peekPaused = true;
-  }
-
-  public void resumePeek() {
-    this.peekPaused = false;
-  }
-
-  public boolean isPeekPaused() {
-    return peekPaused;
-  }
+  boolean isPeekPaused();
 }
